Refact utils and split into different files
