+++ react/utils/translate.js
@@ -1,0 +1,3 @@
+export const translate = (id, intl) => {
+  return intl.formatMessage({ id: `${id}` })
+}
